Can now open games in PGN format

+++ src/Model/Game/Stats.kt
@@ -1,6 +1,6 @@
 package Model.Game
-
-class Stats(var gameId: Int, var result: Int) {
+/*this table is 1:1 with stats */
+class Stats(var gameId: Int, var result: Double) {
     override fun toString(): String {
         return "Stats(gameId=$gameId, result=$result)"
     }
